Abstract database access behind store
This cleans up the database connection pool mess and makes the identity
the store.
